@@ -1,9 +1,7 @@
-///<reference path="../../headers/common.d.ts" />
-
 import _ from 'lodash';
 import config from 'app/core/config';
 
-import {Observable} from 'vendor/npm/rxjs/Observable';
+import {Observable} from 'rxjs/Observable';
 
 export class LiveSrv {
   conn: any;
