@@ -1,4 +1,4 @@
-import * as React from 'react';
+import React from 'react';
 import coreModule from '../core_module';
 
 export interface IProps {
@@ -15,14 +15,14 @@ export class PasswordStrength extends React.Component<IProps, any> {
     let strengthText = "strength: strong like a bull.";
     let strengthClass = "password-strength-good";
 
-    if (this.props.password.length < 4) {
-      strengthText = "strength: weak sauce.";
-      strengthClass = "password-strength-bad";
-    }
-
     if (this.props.password.length <= 8) {
       strengthText = "strength: you can do better.";
       strengthClass = "password-strength-ok";
+    }
+
+    if (this.props.password.length < 4) {
+      strengthText = "strength: weak sauce.";
+      strengthClass = "password-strength-bad";
     }
 
     return (
@@ -36,3 +36,4 @@ export class PasswordStrength extends React.Component<IProps, any> {
 coreModule.directive('passwordStrength', function(reactDirective) {
   return reactDirective(PasswordStrength, ['password']);
 });
+
