@@ -1,6 +1,3 @@
-///<reference path="../headers/common.d.ts" />
-///<reference path="./mod_defs.d.ts" />
-
 import "./directives/dash_class";
 import "./directives/confirm_click";
 import "./directives/dash_edit_link";
@@ -11,7 +8,6 @@ import "./directives/ng_model_on_blur";
 import "./directives/spectrum_picker";
 import "./directives/tags";
 import "./directives/value_select_dropdown";
-import "./directives/plugin_component";
 import "./directives/rebuild_on_change";
 import "./directives/give_focus";
 import "./directives/diff-view";
@@ -20,12 +16,13 @@ import './partials';
 import './components/jsontree/jsontree';
 import './components/code_editor/code_editor';
 import './utils/outline';
+import './components/colorpicker/ColorPicker';
+import './components/colorpicker/SeriesColorPicker';
 
 import {grafanaAppDirective} from './components/grafana_app';
 import {sideMenuDirective} from './components/sidemenu/sidemenu';
 import {searchDirective} from './components/search/search';
 import {infoPopover} from './components/info_popover';
-import {colorPicker} from './components/colorpicker';
 import {navbarDirective} from './components/navbar/navbar';
 import {arrayJoin} from './directives/array_join';
 import {liveSrv} from './live/live_srv';
@@ -57,7 +54,6 @@ export {
   sideMenuDirective,
   navbarDirective,
   searchDirective,
-  colorPicker,
   liveSrv,
   layoutSelector,
   switchDirective,
