@@ -1,4 +1,4 @@
-import {describe, beforeEach, it, expect, angularMocks} from 'test/lib/common';
+import {describe, beforeEach, it, expect, sinon, angularMocks} from 'test/lib/common';
 
 import helpers from 'test/specs/helpers';
 import '../time_srv';
@@ -8,6 +8,7 @@ describe('timeSrv', function() {
   var ctx = new helpers.ServiceTestContext();
   var _dashboard: any = {
     time: {from: 'now-6h', to: 'now'},
+    getTimezone: sinon.stub().returns('browser')
   };
 
   beforeEach(angularMocks.module('grafana.core'));
