@@ -1,3 +1,3 @@
 declare var CloudWatchDatasource: any;
-export {CloudWatchDatasource};
+export default CloudWatchDatasource;
 
