@@ -1,7 +1,7 @@
 
 import {describe, it, expect} from 'test/lib/common';
 
-import queryDef from '../query_def';
+import * as queryDef from '../query_def';
 
 describe('ElasticQueryDef', function() {
 
