@@ -1,3 +1,3 @@
 declare var OpenTsDatasource: any;
-export {OpenTsDatasource};
+export default OpenTsDatasource;
 
