@@ -1,4 +1,4 @@
-import {OpenTsDatasource} from './datasource';
+import OpenTsDatasource from './datasource';
 import {OpenTsQueryCtrl} from './query_ctrl';
 import {OpenTsConfigCtrl} from './config_ctrl';
 
