@@ -1,6 +1,6 @@
 import {describe, beforeEach, it, expect, angularMocks} from 'test/lib/common';
 import helpers from 'test/specs/helpers';
-import {OpenTsDatasource} from "../datasource";
+import OpenTsDatasource from "../datasource";
 
 describe('opentsdb', function() {
   var ctx = new helpers.ServiceTestContext();
