@@ -2,8 +2,6 @@ define([
   'angular',
   'lodash',
   'jquery',
-  'jquery.flot',
-  'jquery.flot.time',
 ],
 function (angular, _, $) {
   'use strict';
@@ -47,7 +45,8 @@ function (angular, _, $) {
             popoverSrv.show({
               element: el[0],
               position: 'bottom center',
-              template: '<gf-color-picker></gf-color-picker>',
+              template: '<series-color-picker series="series" onToggleAxis="toggleAxis" onColorChange="colorSelected">' +
+                '</series-color-picker>',
               openOn: 'hover',
               model: {
                 series: series,
